Generate SDK using latest spec (#367)
* generate SDK over latest spec

* update method signature in types test

* update bridge method name for integration tests

* update method signatures in integration tests

* convert integration tests

* HttpFile fix

* remove generated-client

* update readme

* 7.0.0-rc2

+++ test/it/authserver-claim.ts
@@ -1,16 +1,15 @@
 import { expect } from 'chai';
 import {
   AuthorizationServer,
+  Client,
   Collection,
   DefaultRequestExecutor,
   OAuth2Claim,
   OAuth2Scope
 } from '@okta/okta-sdk-nodejs';
-import type { GeneratedApiClient as V2Client } from '../../src/types/generated-client';
 import getMockAuthorizationServer = require('./mocks/authorization-server');
 import mockScope = require('./mocks/scope.json');
 import mockClaim = require('./mocks/claim.json');
-import utils = require('../utils');
 
 let orgUrl = process.env.OKTA_CLIENT_ORGURL;
 
@@ -18,7 +17,7 @@ if (process.env.OKTA_USE_MOCK) {
   orgUrl = `${orgUrl}/authserver-claim`;
 }
 
-const client: V2Client = utils.getV2Client({
+const client = new Client({
   orgUrl: orgUrl,
   token: process.env.OKTA_CLIENT_TOKEN,
   requestExecutor: new DefaultRequestExecutor()
@@ -28,27 +27,27 @@ describe('Authorization Server Claim API', () => {
   let authServer: AuthorizationServer;
   let scope: OAuth2Scope;
   before(async () => {
-    authServer = await client.createAuthorizationServer(getMockAuthorizationServer());
-    scope = await client.createOAuth2Scope(authServer.id, mockScope as OAuth2Scope);
+    authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: getMockAuthorizationServer()});
+    scope = await client.authorizationServerApi.createOAuth2Scope({authServerId: authServer.id, oAuth2Scope: mockScope as OAuth2Scope});
     expect(scope?.id).to.not.be.undefined;
   });
   after(async () => {
-    await client.deleteOAuth2Scope(authServer.id, scope.id);
-    await client.deleteAuthorizationServer(authServer.id);
+    await client.authorizationServerApi.deleteOAuth2Scope({authServerId: authServer.id, scopeId: scope.id});
+    await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
   });
 
   describe('List all claims', () => {
     let claim: OAuth2Claim;
     beforeEach(async () => {
-      claim = await client.createOAuth2Claim(authServer.id, mockClaim as OAuth2Claim);
+      claim = await client.authorizationServerApi.createOAuth2Claim({authServerId: authServer.id, oAuth2Claim: mockClaim as OAuth2Claim});
       expect(claim?.id).to.not.be.undefined;
     });
     afterEach(async () => {
-      await client.deleteOAuth2Claim(authServer.id, claim.id);
+      await client.authorizationServerApi.deleteOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
     });
 
     it('should return a collection of policies', async () => {
-      const collection = await client.listOAuth2Claims(authServer.id);
+      const collection = await client.authorizationServerApi.listOAuth2Claims({authServerId: authServer.id});
       expect(collection).to.be.instanceOf(Collection);
       const claims = [];
       await collection.each(c => claims.push(c));
@@ -62,11 +61,11 @@ describe('Authorization Server Claim API', () => {
   describe('Create a claim', () => {
     let claim: OAuth2Claim;
     afterEach(async () => {
-      await client.deleteOAuth2Claim(authServer.id, claim.id);
+      await client.authorizationServerApi.deleteOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
     });
 
     it('should get claim from auth server with created claim id', async () => {
-      claim = await client.createOAuth2Claim(authServer.id, mockClaim as OAuth2Claim);
+      claim = await client.authorizationServerApi.createOAuth2Claim({authServerId: authServer.id, oAuth2Claim: mockClaim as OAuth2Claim});
       expect(claim).to.be.exist;
       expect(claim.name).to.equal(mockClaim.name);
     });
@@ -75,14 +74,14 @@ describe('Authorization Server Claim API', () => {
   describe('Get a claim', () => {
     let claim: OAuth2Claim;
     beforeEach(async () => {
-      claim = await client.createOAuth2Claim(authServer.id, mockClaim as OAuth2Claim);
+      claim = await client.authorizationServerApi.createOAuth2Claim({authServerId: authServer.id, oAuth2Claim: mockClaim as OAuth2Claim});
     });
     afterEach(async () => {
-      await client.deleteOAuth2Claim(authServer.id, claim.id);
+      await client.authorizationServerApi.deleteOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
     });
 
     it('should get claim from auth server by id', async () => {
-      const claimFromGet = await client.getOAuth2Claim(authServer.id, claim.id);
+      const claimFromGet = await client.authorizationServerApi.getOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
       expect(claimFromGet).to.be.instanceOf(OAuth2Claim);
       expect(claimFromGet.id).to.equal(claim.id);
     });
@@ -91,16 +90,16 @@ describe('Authorization Server Claim API', () => {
   describe('Update claim', () => {
     let claim: OAuth2Claim;
     beforeEach(async () => {
-      claim = await client.createOAuth2Claim(authServer.id, mockClaim as OAuth2Claim);
+      claim = await client.authorizationServerApi.createOAuth2Claim({authServerId: authServer.id, oAuth2Claim: mockClaim as OAuth2Claim});
     });
     afterEach(async () => {
-      await client.deleteOAuth2Claim(authServer.id, claim.id);
+      await client.authorizationServerApi.deleteOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
     });
 
     it('should update name for created scope', async () => {
       const mockName = 'MockUpdateClaim';
       claim.name = mockName;
-      const updatedClaim = await client.updateOAuth2Claim(authServer.id, claim.id, claim);
+      const updatedClaim = await client.authorizationServerApi.replaceOAuth2Claim({authServerId: authServer.id, claimId: claim.id, oAuth2Claim: claim});
       expect(updatedClaim.id).to.equal(claim.id);
       expect(updatedClaim.name).to.equal(mockName);
     });
@@ -109,14 +108,14 @@ describe('Authorization Server Claim API', () => {
   describe('Delete claim', () => {
     let claim: OAuth2Claim;
     beforeEach(async () => {
-      claim = await client.createOAuth2Claim(authServer.id, mockClaim as OAuth2Claim);
+      claim = await client.authorizationServerApi.createOAuth2Claim({authServerId: authServer.id, oAuth2Claim: mockClaim as OAuth2Claim});
     });
 
     it('should not get claim after deletion', async () => {
-      const res = await client.deleteOAuth2Claim(authServer.id, claim.id);
+      const res = await client.authorizationServerApi.deleteOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
       expect(res).to.equal(undefined);
       try {
-        await client.getOAuth2Claim(authServer.id, claim.id);
+        await client.authorizationServerApi.getOAuth2Claim({authServerId: authServer.id, claimId: claim.id});
       } catch (e) {
         expect(e.status).to.equal(404);
       }
