+++ test/it/authserver-credential.ts
@@ -1,13 +1,12 @@
 import { expect } from 'chai';
 import {
   AuthorizationServer,
+  Client,
   Collection,
   DefaultRequestExecutor,
   JsonWebKey,
 } from '@okta/okta-sdk-nodejs';
 import getMockAuthorizationServer = require('./mocks/authorization-server');
-import type { GeneratedApiClient as V2Client } from '../../src/types/generated-client';
-import utils = require('../utils');
 
 let orgUrl = process.env.OKTA_CLIENT_ORGURL;
 
@@ -15,7 +14,7 @@ if (process.env.OKTA_USE_MOCK) {
   orgUrl = `${orgUrl}/authserver-credential`;
 }
 
-const client: V2Client = utils.getV2Client({
+const client = new Client({
   orgUrl: orgUrl,
   token: process.env.OKTA_CLIENT_TOKEN,
   requestExecutor: new DefaultRequestExecutor()
@@ -24,15 +23,15 @@ const client: V2Client = utils.getV2Client({
 describe('Authorization Server Credential API', () => {
   let authServer: AuthorizationServer;
   before(async () => {
-    authServer = await client.createAuthorizationServer(getMockAuthorizationServer());
+    authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: getMockAuthorizationServer()});
   });
   after(async () => {
-    await client.deleteAuthorizationServer(authServer.id);
+    await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
   });
 
   describe('Get Authorization Server Keys', () => {
     it('should return a collection of JsonWebKey', async () => {
-      const collection = await client.listAuthorizationServerKeys(authServer.id);
+      const collection = await client.authorizationServerApi.listAuthorizationServerKeys({authServerId: authServer.id});
       expect(collection).to.be.instanceOf(Collection);
       await collection.each(key => {
         expect(key).to.be.instanceOf(JsonWebKey);
@@ -42,7 +41,7 @@ describe('Authorization Server Credential API', () => {
 
   describe('Rotate Authorization Server Keys', () => {
     it('should return a collection of JsonWebKey', async () => {
-      const collection = await client.rotateAuthorizationServerKeys(authServer.id, { use: 'sig' });
+      const collection = await client.authorizationServerApi.rotateAuthorizationServerKeys({authServerId: authServer.id, use: { use: 'sig' }});
       expect(collection).to.be.instanceOf(Collection);
       await collection.each(key => {
         expect(key).to.be.instanceOf(JsonWebKey);
