+++ test/it/authserver-crud.ts
@@ -2,12 +2,11 @@ import { expect } from 'chai';
 import { spy } from 'sinon';
 import {
   AuthorizationServer,
+  Client,
   DefaultRequestExecutor
 } from '@okta/okta-sdk-nodejs';
-import type { GeneratedApiClient as V2Client } from '../../src/types/generated-client';
 import getMockAuthorizationServer = require('./mocks/authorization-server');
 import faker = require('@faker-js/faker');
-import utils = require('../utils');
 
 let orgUrl = process.env.OKTA_CLIENT_ORGURL;
 
@@ -15,7 +14,7 @@ if (process.env.OKTA_USE_MOCK) {
   orgUrl = `${orgUrl}/authserver-crud`;
 }
 
-const client: V2Client = utils.getV2Client({
+const client = new Client({
   orgUrl: orgUrl,
   token: process.env.OKTA_CLIENT_TOKEN,
   requestExecutor: new DefaultRequestExecutor()
@@ -25,13 +24,13 @@ describe('Authorization Server Crud API', () => {
   describe('Create Auth Server', () => {
     let authServer: AuthorizationServer;
     afterEach(async () => {
-      await client.deactivateAuthorizationServer(authServer.id);
-      await client.deleteAuthorizationServer(authServer.id);
+      await client.authorizationServerApi.deactivateAuthorizationServer({authServerId: authServer.id});
+      await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
     });
 
     it('should return correct model', async () => {
       const mockAuthorizationServer = getMockAuthorizationServer();
-      authServer = await client.createAuthorizationServer(mockAuthorizationServer);
+      authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: mockAuthorizationServer});
       expect(authServer).to.be.instanceOf(AuthorizationServer);
       expect(authServer.id).to.be.exist;
       expect(authServer.name).to.be.equal(mockAuthorizationServer.name);
@@ -52,7 +51,7 @@ describe('Authorization Server Crud API', () => {
             ...getMockAuthorizationServer(),
             name: `node-sdk: ${prefix} ${i} ${faker.random.word()}`.substring(0, 49)
           };
-          const createdServer = await client.createAuthorizationServer(newServer);
+          const createdServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: newServer});
           createdServers.push(createdServer);
         }
       }
@@ -60,19 +59,18 @@ describe('Authorization Server Crud API', () => {
     });
     after(async () => {
       for (const authServer of createdServers) {
-        await client.deactivateAuthorizationServer(authServer.id);
-        await client.deleteAuthorizationServer(authServer.id);
+        await client.authorizationServerApi.deactivateAuthorizationServer({authServerId: authServer.id});
+        await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
       }
     });
 
     it('should search with q and paginate results', async () => {
       const queryParameters = {
         q: 'node-sdk: AUTH_SRV_AB',
-        // TODO: OKTA-515269 - limit should be number
-        limit: '1'
+        limit: 1
       };
       const filtered = new Set();
-      const collection = await client.listAuthorizationServers(queryParameters);
+      const collection = await client.authorizationServerApi.listAuthorizationServers({...queryParameters});
       const pageSpy = spy(collection, 'getNextPage');
       await collection.each(as => {
         expect(as).to.be.an.instanceof(AuthorizationServer);
@@ -88,16 +86,16 @@ describe('Authorization Server Crud API', () => {
   describe('Get Authorization Server', () => {
     let authServer: AuthorizationServer;
     beforeEach(async () => {
-      authServer = await client.createAuthorizationServer(getMockAuthorizationServer());
+      authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: getMockAuthorizationServer()});
     });
 
     afterEach(async () => {
-      await client.deactivateAuthorizationServer(authServer.id);
-      await client.deleteAuthorizationServer(authServer.id);
+      await client.authorizationServerApi.deactivateAuthorizationServer({authServerId: authServer.id});
+      await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
     });
 
     it('should get Authorization Server by id', async () => {
-      const authServerFromGet = await client.getAuthorizationServer(authServer.id);
+      const authServerFromGet = await client.authorizationServerApi.getAuthorizationServer({authServerId: authServer.id});
       expect(authServerFromGet).to.be.instanceOf(AuthorizationServer);
       expect(authServerFromGet.name).to.equal(authServer.name);
     });
@@ -106,17 +104,17 @@ describe('Authorization Server Crud API', () => {
   describe('Update Authorization Server', () => {
     let authServer: AuthorizationServer;
     beforeEach(async () => {
-      authServer = await client.createAuthorizationServer(getMockAuthorizationServer());
+      authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: getMockAuthorizationServer()});
     });
     afterEach(async () => {
-      await client.deactivateAuthorizationServer(authServer.id);
-      await client.deleteAuthorizationServer(authServer.id);
+      await client.authorizationServerApi.deactivateAuthorizationServer({authServerId: authServer.id});
+      await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
     });
 
     it('should update name for created auth server', async () => {
       const mockName = 'Mock update auth server';
       authServer.name = mockName;
-      const updatedAuthServer = await client.updateAuthorizationServer(authServer.id, authServer);
+      const updatedAuthServer = await client.authorizationServerApi.replaceAuthorizationServer({authServerId: authServer.id, authorizationServer: authServer});
       expect(updatedAuthServer.id).to.equal(authServer.id);
       expect(updatedAuthServer.name).to.equal(mockName);
     });
@@ -125,15 +123,15 @@ describe('Authorization Server Crud API', () => {
   describe('Delete Authorization Server', () => {
     let authServer: AuthorizationServer;
     beforeEach(async () => {
-      authServer = await client.createAuthorizationServer(getMockAuthorizationServer());
+      authServer = await client.authorizationServerApi.createAuthorizationServer({authorizationServer: getMockAuthorizationServer()});
     });
 
     it('should not get authserver after deletion', async () => {
-      await client.deactivateAuthorizationServer(authServer.id);
-      const res = await client.deleteAuthorizationServer(authServer.id);
+      await client.authorizationServerApi.deactivateAuthorizationServer({authServerId: authServer.id});
+      const res = await client.authorizationServerApi.deleteAuthorizationServer({authServerId: authServer.id});
       expect(res).to.equal(undefined);
       try {
-        await client.getAuthorizationServer(authServer.id);
+        await client.authorizationServerApi.getAuthorizationServer({authServerId: authServer.id});
       } catch (e) {
         expect(e.status).to.equal(404);
       }
